requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
dice-ml==0.12
lime==0.2.0.1
matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
PySMT==0.9.6
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.0
torch==2.10.0+cu126
torchvision==0.25.0+cu126
tqdm==4.67.2
z3-solver==4.16.0.0
plotly>=5.20
streamlit>=1.32
umap-learn>=0.5
pymoo>=0.6.0
psutil>=5.9
pyarrow>=14.0
streamlit-plotly-events>=0.0.6
